Fix Stack link (#6)
Without the `https://` in the beginning, Github turns urls into relative links inside the Github project, for some reason. Instead of going to the indented `haskellstack.org`, it goes to `https://github.com/NorfairKing/smos/blob/master/haskellstack.org`.

Replace with a version of the URL that includes `https://`. Also use the canonical URL, to avoid a redirect.