Don't Set TODO State if Template State Set to null
Currently, using `state: null` in a template will be overriden,
resulting in a `state: TODO` in the generated file.

This commit changes this behavior, only setting the generated state to
`TODO` if the `state` field is missing. Instead of parsing a `Maybe
TodoState` in the template, we parse a `Maybe (Maybe TodoState)`. The
`Nothing` case indicates the field does not exist, while the `Just
Nothing` indicates it was explicitly set to `null`.